removed whats new md and moved to folder
