--- src/neuroglancer/whats_new.md
@@ -1,2 +1,0 @@
-#### Hello World
-Nothing to see here folks...
