+++ src/neuroglancer/whats_new/whats_new.css
